requirements.txt
PyQt5>=5.15.0
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
torch>=1.7.0
git+https://github.com/openai/CLIP.git
imagehash>=4.2.0
scikit-learn>=0.24.0
